% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fspec-genomic.R
\name{bed}
\alias{bed}
\title{BED files.}
\format{Tab-delimited file genomic interval format.}
\usage{
bed()
}
\description{
see \url{https://genome.ucsc.edu/FAQ/FAQformat.html} for format details.
N.B. chrom, chromStart, chromEnd != seqid, start, end.
}

