% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.R
\name{complete_spec}
\alias{complete_spec}
\title{complete a file specification with any unset defaults}
\usage{
complete_spec(file_spec)
}
\arguments{
\item{file_spec}{(named list) forming a file specification.}
}
\description{
complete a file specification with any unset defaults
}

