% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/input_list.R
\name{do_read}
\alias{do_read}
\title{read delimited file with using a named list of arguments.}
\usage{
do_read(.il)
}
\arguments{
\item{.il}{input_list. see \code{skel_*} functions.}
}
\description{
read delimited file with using a named list of arguments.
}
\keyword{internal}

