% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fspec.R
\name{fspec}
\alias{fspec}
\title{Construct a file specification.}
\usage{
fspec(class, .fspec, .l)
}
\arguments{
\item{class}{(chr) Assign a class to the file specification.}

\item{.fspec}{(fspec object) Optionally, fpsec object to inherit from.}

\item{.l}{(named list) Provide a list of parameters to serve as defaults.}
}
\description{
A file specification (fspec) defines the set of consistent arguments to be
used to read in files of a given type. It should be a named list with
elements corresponding to the arguments (formals) of the function used to
read it in.
}

