% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_first_line}
\alias{get_first_line}
\title{get first data line of file}
\usage{
get_first_line(file, delim)
}
\arguments{
\item{file}{path to a file.}

\item{delim}{delimiter of data records in file. passes to: \code{\link{
tokenizer_delim}}}
}
\description{
get first data line of file
}
\keyword{internal}

