% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fspec-genomic.R
\name{gff}
\alias{gff}
\title{GFF files}
\format{Tab-delimited genomic interval file.}
\usage{
gff()
}
\description{
see \url{http://gmod.org/wiki/GFF3} for format details. does not handle fasta
containing files. For GFF2 spec see:
\url{http://www.sanger.ac.uk/resources/software/gff/spec.html#t_2}.
}

