% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{infer_data_records}
\alias{infer_data_records}
\title{infer data records by expected fields}
\usage{
infer_data_records(file, delim, n_fields)
}
\arguments{
\item{file}{path to file}

\item{delim}{delimiter of data records in file}

\item{n_fields}{number of \code{delim} fields (cols) to expect}
}
\description{
infer data records by expected fields
}
\keyword{internal}

