% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/infer_.R
\name{infer_seqid}
\alias{infer_seqid}
\title{Infer the name of the sequence, chromosome, or scaffold.}
\usage{
infer_seqid(tbl, col = NULL)
}
\arguments{
\item{tbl}{data.frame to pass in}

\item{col}{the column to infer and mutate if necessary.}
}
\description{
the name of a chromosome should not contain "chr". this removes it if present
}

