% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ls_read.R
\name{ls_read}
\alias{ls_read}
\title{Read a list of files of the same format based on a pattern and path.
Optionally bind them together, adding a variable containing the filename
for each observation.}
\usage{
ls_read(ptrn = c(".gff3", ".wig"), path = ".", bind_rows = T)
}
\arguments{
\item{ptrn}{character. pattern by which to look up files. defaults to gff3.}

\item{path}{character. path to directory to look for files in.
defaults to current.}

\item{bind_rows}{logical. should all the resulting dfs be bound into one?}
}
\description{
Read a list of files of the same format based on a pattern and path.
Optionally bind them together, adding a variable containing the filename
for each observation.
}
\examples{
\dontrun{
 ls_read(".gff3", "./data-raw")
 }
}

