% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bed.R
\name{read_bed}
\alias{read_bed}
\title{read in a bed file}
\usage{
read_bed(file, type = "bed6")
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data
   (either a single string or a raw vector).

   Files ending in \code{.gz}, \code{.bz2}, \code{.xz}, or \code{.zip} will
   be automatically uncompressed. Files starting with \code{http://},
   \code{https://}, \code{ftp://}, or \code{ftps://} will be automatically
   downloaded.

   Literal data is most useful for examples and tests. It must contain at
   least one new line to be recognised as data (instead of a path).}

\item{type}{The type of bed file. One of
\code{c("bed3", "bed4", "bed6", "bed12")}.}
}
\value{
3, 4, 6, or 12 variable tbl_df and data.frame.
}
\description{
see \url{https://genome.ucsc.edu/FAQ/FAQformat.html} for format details.
N.B. chrom, chromStart, chromEnd != seqid, start, end.
}
\examples{
\dontrun{
 # literal string
 st <- "chr3R\\t2\\t37\\tyfg\\t0.13859536\\t+\\n"
 read_bed(st)

 # local file, may be compressed
 read_bed("data-raw/three_col.bed.gz", type = "bed3")
 read_bed("data-raw/three_col.bed.gz")

 # or you could borow an internet
 }
}

