% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biogrid.R
\name{read_biogrid}
\alias{read_biogrid}
\title{read in a BioGRID TAB 2.0 file}
\usage{
read_biogrid(file, type = "tab2i")
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data
   (either a single string or a raw vector).

   Files ending in \code{.gz}, \code{.bz2}, \code{.xz}, or \code{.zip} will
   be automatically uncompressed. Files starting with \code{http://},
   \code{https://}, \code{ftp://}, or \code{ftps://} will be automatically
   downloaded.

   Literal data is most useful for examples and tests. It must contain at
   least one new line to be recognised as data (instead of a path).}

\item{type}{The type of bed file. One of \code{c("tab2i", "tab2c")}.}
}
\value{
24 or 23 variable tbl_df and data.frame.
}
\description{
see \url{http://wiki.thebiogrid.org/doku.php/biogrid_tab_version_2.0} for
format details.
}

