% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.R
\name{read_fspec}
\alias{read_fspec}
\title{Read a delimited file from a file specification}
\usage{
read_fspec(file, .fspec, .f = readr::read_delim, infr_skip = TRUE)
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data
   (either a single string or a raw vector).

   Files ending in \code{.gz}, \code{.bz2}, \code{.xz}, or \code{.zip} will
   be automatically uncompressed. Files starting with \code{http://},
   \code{https://}, \code{ftp://}, or \code{ftps://} will be automatically
   downloaded.

   Literal data is most useful for examples and tests. It must contain at
   least one new line to be recognised as data (instead of a path).}

\item{.fspec}{(fspec object) Optionally, fpsec object to inherit from.}
}
\description{
Read a delimited file from a file specification
}
\examples{
\dontrun{
 read_fspec(file = "data-raw/three_col.bed.gz", bed_bed3)
 }
}

