% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gff3.R
\name{read_gff3}
\alias{read_gff3}
\title{read in a gff3 file}
\usage{
read_gff3(file)
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data
   (either a single string or a raw vector).

   Files ending in \code{.gz}, \code{.bz2}, \code{.xz}, or \code{.zip} will
   be automatically uncompressed. Files starting with \code{http://},
   \code{https://}, \code{ftp://}, or \code{ftps://} will be automatically
   downloaded.

   Literal data is most useful for examples and tests. It must contain at
   least one new line to be recognised as data (instead of a path).}
}
\value{
9 variable tbl_df and data.frame.
}
\description{
see \url{http://gmod.org/wiki/GFF3} for format details. does not handle fasta
containing files.
}
\examples{
# literal string
 st <- "3R\\treg\\tbind_site\\t46748\\t48137\\t0.499\\t.\\t.\\tID=enr_reg_1\\n"
 read_gff3(st)
 \dontrun{
 # local file, may be compressed
 read_gff3("../data-raw/file.gff3")
 read_gff3("../data-raw/file.gff3.gz")

 # or you could borow an internet
 }
}

