% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gff3.R
\name{read_wig}
\alias{read_wig}
\title{read in a wig file}
\usage{
read_wig(file, col_names = c("seqid", "start", "end", "score"),
  col_types = "cddd", na = "NA", progress = FALSE, delim = " ",
  skip = 1, ...)
}
\arguments{
\item{file}{path to a file.}

\item{col_names}{passed to \code{\link{read_delim}}'s col_names}

\item{col_types}{passed to \code{\link{read_delim}}'s col_types}

\item{na}{passed to \code{\link{read_delim}}'s na}

\item{progress}{passed to \code{\link{read_delim}}'s progress}

\item{delim}{passed to \code{\link{read_delim}}'s delim. These can be space
or tab delimited so need to test.}

\item{other}{dots you wish to pass to \code{\link{read_delim}}}
}
\description{
see \url{http://gmod.org/wiki/GBrowse/Uploading_Wiggle_Tracks#Format_Descriptions}
for format details.
Wraps \code{\link{read_tsv}} and guesses how many lines to skip.
}
\details{
N.B. currently only handles wiggle "BED format" type.
}
\examples{
\dontrun{
 tmp <- "data-raw/wiggle_bedtype.wig.gz"
 read_gff3(tmp)
 }
}

