% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zzz.r
\docType{package}
\name{readrbio}
\alias{readrbio}
\alias{readrbio-package}
\title{Convenience functions for reading biological file formats.}
\description{
Convenience functions for reading biological file formats.
}
\section{Formats of interest}{

\itemize{
 \item GFF3
 \item WIG
}
}

