% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wig.R
\name{skel_wig}
\alias{skel_wig}
\title{make a skeleton input list for a wig file.}
\usage{
skel_wig(file, type)
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data
   (either a single string or a raw vector).

   Files ending in \code{.gz}, \code{.bz2}, \code{.xz}, or \code{.zip} will
   be automatically uncompressed. Files starting with \code{http://},
   \code{https://}, \code{ftp://}, or \code{ftps://} will be automatically
   downloaded.

   Literal data is most useful for examples and tests. It must contain at
   least one new line to be recognised as data (instead of a path).}

\item{type}{one of c("bed", "fixed", "variable"). N.B. only "bed" currently.}
}
\value{
input_list: a named list of arguments relevant to read_* functions.
}
\description{
make a skeleton input list for a wig file.
}
\keyword{internal}

