% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/input-spec.R
\name{spec}
\alias{spec}
\title{Build a file specification}
\usage{
spec(file, type, subtype)
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data
   (either a single string or a raw vector).

   Files ending in \code{.gz}, \code{.bz2}, \code{.xz}, or \code{.zip} will
   be automatically uncompressed. Files starting with \code{http://},
   \code{https://}, \code{ftp://}, or \code{ftps://} will be automatically
   downloaded.

   Literal data is most useful for examples and tests. It must contain at
   least one new line to be recognised as data (instead of a path).}

\item{type}{(chr) the file type.}

\item{subtype}{(chr) the file subtype.}
}
\description{
Build a file specification
}

