% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spec-biogrid.R
\name{spec_biogrid}
\alias{spec_biogrid}
\title{BioGRID TAB 2.0 files}
\format{Tab-delimited file format from BioGRID.}
\usage{
spec_biogrid()
}
\description{
see \url{http://wiki.thebiogrid.org/doku.php/biogrid_tab_version_2.0} for
format details.
}

