% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spec-wig.R
\name{spec_wig}
\alias{spec_wig}
\title{WIGGLE track files}
\format{space-delimited dense genomic signal file.}
\usage{
spec_wig()
}
\description{
see \url{http://gmod.org/wiki/GBrowse/Uploading_Wiggle_Tracks#Format_Descriptions}
for format details. does not handle fasta containing files.
}

