% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{translate_skip_n_max}
\alias{translate_skip_n_max}
\title{translate vector of data_records to skip and n_max params.}
\usage{
translate_skip_n_max(data_records)
}
\arguments{
\item{data_records}{}
}
\description{
Currently assumes a contiguous data records section
}
\keyword{internal}

