% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fspec-genomic.R
\name{wig}
\alias{wig}
\title{WIGGLE track files}
\format{space-delimited dense genomic signal file.}
\usage{
wig()
}
\description{
see \url{http://gmod.org/wiki/GBrowse} for format details. Does not handle
fasta containing files.
}

