% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/infer_.R
\name{infer_skip}
\alias{infer_skip}
\title{Infer how many rows to skip.}
\usage{
infer_skip(.il)
}
\arguments{
\item{.il}{input_list. see \code{skel_*} functions.}
}
\description{
Infer a skip parameter by comparing the number of fields in the first 100
rows to the number of data fields expected.
}
\keyword{internal}

