% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/infr.R
\name{infr_skip}
\alias{infr_skip}
\title{Infer how many rows to skip.}
\usage{
infr_skip(.fspec, .n = 100L, skip_if_start = "#")
}
\arguments{
\item{.fspec}{(named list) forming a file specification.}

\item{.n}{integer. infer using first \code{.n} lines.}

\item{skip_if_start}{character. if line starts with this ignore it.}
}
\description{
Infer a skip parameter by comparing the number of fields in the first n
rows to the number of data fields expected. It also takes into account
any rows that begin with a comment-type character; ignoring rows starting
with \code{skip_if_start}
}

