% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gff3.R
\name{read_gff3}
\alias{read_gff3}
\title{read in a gff3 file}
\usage{
read_gff3(file, col_names = c("seqid", "source", "type", "start", "end",
  "score", "strand", "phase", "attributes"), col_types = "cccdddccc",
  na = ".", progress = FALSE, ...)
}
\arguments{
\item{file}{path to a file.}

\item{col_names}{passed to \code{\link{read_tsv}}'s col_names}

\item{col_types}{passed to \code{\link{read_tsv}}'s col_types}

\item{na}{passed to \code{\link{read_tsv}}'s na}

\item{progress}{passed to \code{\link{read_tsv}}'s na}

\item{other}{dots you wish to pass to \code{\link{read_tsv}}}
}
\description{
see \url{http://gmod.org/wiki/GFF3} for format details.
Wraps \code{\link{read_tsv}} and guesses how many lines to skip.
}
\examples{
\dontrun{
 tmp <- "data-raw/file.gff3.gz"
 read_gff3(tmp)
 }
}

