% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wig.R
\name{read_wig}
\alias{read_wig}
\title{read in a wig file}
\usage{
read_wig(file, type = "bed")
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data
   (either a single string or a raw vector).

   Files ending in \code{.gz}, \code{.bz2}, \code{.xz}, or \code{.zip} will
   be automatically uncompressed. Files starting with \code{http://},
   \code{https://}, \code{ftp://}, or \code{ftps://} will be automatically
   downloaded.

   Literal data is most useful for examples and tests. It must contain at
   least one new line to be recognised as data (instead of a path).}

\item{type}{one of c("bed", "fixed", "variable"). N.B. only "bed" currently.}
}
\value{
4 variable tbl_df and data.frame.
}
\description{
see \url{http://gmod.org/wiki/GBrowse/Uploading_Wiggle_Tracks#Format_Descriptions}
for format details. does not handle fasta containing files.
}
\examples{
# literal string
 st <- "chr3R 2 37 0.13859536\\n"
 read_wig(st)
 \dontrun{
 # local file, may be compressed
 read_wig("data-raw/wiggle_bedtype.wig")
 read_wig("data-raw/wiggle_bedtype.wig.gz")

 # or you could borow an internet
 }
}

